pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-fraud"
version = "0.1.0"
requires-python = ">=3.10"
keywords = [
    "fraud detection",
    "machine learning",
    "ml pipeline",
    "anomaly detection",
    "classification",
    "data science",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "shap>=0.42.0",
    "imbalanced-learn>=0.11.0",
    "PyYAML>=6.0",
    "joblib>=1.3.0",
    "ensure>=0.10.0",
    "python-box>=0.4.1",
    "ensure>=0.10.0",
    "python-box>=0.4.1",
    "optuna>=3.0.0",
    "mlflow>=3.10.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.2.0",
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.3.1",
]

[project.urls]
Homepage = "https://github.com/minich-code/financial-fraud"
"Bug Tracker" = "https://github.com/minich-code/financial-fraud/issues"

[tool.hatch.build]
include = [
    "src/**/*.py",
    "src/**/*.json",
    "src/**/*.yml",
    "src/**/*.yaml",
    "config/**/*.yaml",
    "LICENSE",
    "README.md",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "tests/**",
    "docs/**",
    "artifacts/**",
    ".env*",
    "*.log",
    "logs/**",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3